Create NovelScrapper.py
WebScrapping Online Novels: The project will parse the table of contents of the specified novel from a website and scrap the data of each chapter of the novel and store the extracted data in local files. The project is implemented using Python using libraries such as Beautiful Soap and Requests module.